Improve startupscript, it takes stdin and gives a nice error msg

+++ production/errantia_prod.sh
@@ -1,7 +1,18 @@
 #!/bin/bash
 
+if [ -t 0 ]; then
+    echo
+    echo "No pipe detected. You need to pipe a file or a command. E.g.:"
+    echo " dvgrab | $0"
+    echo " cat video.dv | $0"
+    echo
+
+    exit 1
+fi
+
 pkill -f "python sub.py"
-python sub.py &
+# Redirect stdin into the sub.py python script (should be a DV stream (or file))
+python sub.py <&0 &
 
 sleep 1
